+++ packages/ubq/CHANGELOG.md
@@ -15,6 +15,19 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 
+## [5.23.18] - 2022-09-12
+### Fixed
+ - use exactly `cron-parser@4.4.0` (last version what uses require) instead of `^4.4.0`
+
+## [5.23.17] - 2022-09-11
+### Added
+ - new method `ubq_scheduler.calculateNextCronTerm` - return next `cnt` execution dates for specified cronExpression;
+   Supports non-standard 7-part cron expression syntax, where last field `@occurrence` mean - fires on every x occurrence.
+   Example - `0 0 1 * * 1 @2` = At 01:00 AM, only on Monday, once per 2 occurrence (every second Monday)
+
+ - new entity-level method `ubq_scheduler.estimateCronSchedule` - wrapper for `ubq_scheduler.calculateNextCronTerm`
+  for client-side execution. Used by UCron to show "Estimated executions"
+
 ## [5.23.16] - 2022-08-26
 ## [5.23.15] - 2022-08-23
 ## [5.23.14] - 2022-08-19